+++ manticore/native/cpu/cpufactory.py
