+++ manticore/native/manticore.py
@@ -117,7 +117,7 @@ def _make_initial_state(binary_path, **kwargs):
 
 
 def _make_decree(program, concrete_start='', **kwargs):
-    from manticore.platforms import decree
+    from ..platforms import decree
 
     constraints = ConstraintSet()
     platform = decree.SDecree(constraints, program)
@@ -132,7 +132,7 @@ def _make_decree(program, concrete_start='', **kwargs):
 
 
 def _make_linux(program, argv=None, env=None, entry_symbol=None, symbolic_files=None, concrete_start='', stdin_size=consts.stdin_size):
-    from manticore.platforms import linux
+    from ..platforms import linux
 
     env = {} if env is None else env
     argv = [] if argv is None else argv
