+++ manticore/native/models.py
@@ -2,7 +2,7 @@
 Models here are intended to be passed to :meth:`~manticore.native.state.State.invoke_model`, not invoked directly.
 """
 
-from ..core.cpu.abstractcpu import ConcretizeArgument
+from .cpu.abstractcpu import ConcretizeArgument
 from ..utils.helpers import issymbolic
 from ..core.smtlib.solver import solver
 from ..core.smtlib.operators import ITEBV, ZEXTEND
