+++ manticore/platforms/linux.py
@@ -2,6 +2,7 @@ import binascii
 import ctypes
 import errno
 import fcntl
+import logging
 import socket
 from typing import Union, List, TypeVar, cast
 
@@ -14,13 +15,12 @@ from elftools.elf.elffile import ELFFile
 from elftools.elf.sections import SymbolTableSection
 
 from . import linux_syscalls
-from ..core.cpu.abstractcpu import Syscall, ConcretizeArgument
-from ..core.cpu.arm import *
-from ..core.cpu.cpufactory import CpuFactory
 from ..core.executor import TerminateState
 from ..core.smtlib import ConstraintSet, solver
 from ..core.smtlib import Expression
 from ..exceptions import SolverException
+from ..native.cpu.abstractcpu import Syscall, ConcretizeArgument
+from ..native.cpu.cpufactory import CpuFactory
 from ..native.memory import SMemory32, SMemory64, Memory32, Memory64
 from ..platforms.platform import Platform, SyscallNotImplemented
 from ..utils.helpers import issymbolic
