+++ manticore/utils/helpers.py
@@ -19,7 +19,7 @@ def issymbolic(value):
     :return: whether `value` is symbolic
     :rtype: bool
     '''
-    from manticore.core.smtlib import Expression  # prevent circular imports
+    from ..core.smtlib import Expression  # prevent circular imports
     return isinstance(value, Expression)
 
 
@@ -66,7 +66,7 @@ def taint_with(arg, taint, value_bits=256, index_bits=256):
     :param arg: a value or Expression
     :param taint: a regular expression matching a taint value (eg. 'IMPORTANT.*'). If None, this function checks for any taint value.
     '''
-    from manticore.core.smtlib import BitVecConstant  # prevent circular imports
+    from ..core.smtlib import BitVecConstant  # prevent circular imports
 
     tainted_fset = frozenset((taint,))
