+++ tests/test_armv7cpu.py
@@ -5,8 +5,8 @@ from capstone import CS_MODE_THUMB, CS_MODE_ARM
 from functools import wraps
 from keystone import Ks, KS_ARCH_ARM, KS_MODE_ARM, KS_MODE_THUMB
 
-from manticore.core.cpu.abstractcpu import ConcretizeRegister
-from manticore.core.cpu.arm import Armv7Cpu as Cpu, Mask, Interruption
+from manticore.native.cpu import ConcretizeRegister
+from manticore.native.cpu import Armv7Cpu as Cpu, Mask, Interruption
 from manticore.core.smtlib import *
 from manticore.core.state import Concretize
 from manticore.native.memory import SMemory32
