+++ tests/test_cpu_automatic.py
@@ -1,6 +1,5 @@
 
 import unittest
-from manticore.core.cpu.x86 import *
 from manticore.native.memory import *
 
 class CPUTest(unittest.TestCase):
