+++ tests/test_dyn.py
@@ -1,7 +1,5 @@
 
 import unittest
-from manticore.core.cpu.x86 import *
-from manticore.core.smtlib import Operators
 from manticore.native.memory import *
